Banking System program
So I was Asked to make a banking system program that allows a user to create an account and generate a 4 digit pin the user will use to access his/her details and make deposits and withdrawals, I succeeded in most part but then getting the code to ask for the pin and account details was a bit of a hassle for me so i decided to ignore that part and just make a normal one. Please if anyone can help me with this challenge, please do.
This is my code after i learnt OOP in C++ some days before i got this assignment i used 1 week to do.
Well youtube also played some part in my work though.
Anyways yeah......It doesn't matter if i have started poorly. I'LL STILL BE THE BEST!!!